python/clean_trello/requirements.txt
astroid==2.0.4
autopep8==1.4
certifi==2018.8.24
chardet==3.0.4
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
py-trello==0.12.0
pycodestyle==2.4.0
pylint==2.1.1
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
wrapt==1.10.11
